previous/requirements.txt
langchain==0.0.284
python-dotenv==1.0.0
streamlit==1.22.0
unstructured==0.9.2
"unstructured[pdf]"
tiktoken==0.4.0
faiss-cpu==1.7.4
sentence-transformers
libmagic==1.0
python-magic==0.4.27
python-magic-bin==0.4.14
OpenAI == 0.28.0
google-generativeai
langchain-google-genai